Split SvgFilter
